requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-multiselectfield==0.1.12
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1
